requirements.txt
pytorch
torchvision
torchviz